[Diffusion] Isolate HWR startup integration
